lib/balloon_util.c: fix a compiler warning
A warning emitted by gcc-3.4.4 (CentOS 4) on an x86_64:

balloon_util.c: In function `fiemap_get':
balloon_util.c:162: warning: long long unsigned int format, __u64 arg (arg 3)
balloon_util.c:162: warning: long long unsigned int format, __u64 arg (arg 4)
balloon_util.c:162: warning: long long unsigned int format, __u64 arg (arg 5)

+++ lib/balloon_util.c
@@ -157,9 +157,9 @@ int fiemap_get(int fd, __u64 off, __u64 start, off_t size, struct pfiemap **pfie
 			    ~FIEMAP_EXTENT_LAST) {
 				ploop_err(0, "Skipping extent (%llu/%llu"
 					"/%llu) with unexpected flags=%s",
-					fm_ext[i].fe_length,
-					fm_ext[i].fe_logical,
-					fm_ext[i].fe_physical,
+					(long long unsigned)fm_ext[i].fe_length,
+					(long long unsigned)fm_ext[i].fe_logical,
+					(long long unsigned)fm_ext[i].fe_physical,
 					fl(fm_ext[i].fe_flags));
 				continue;
 			}
